requirements.txt
ultralytics
opencv-python
numpy
torch
torchvision
sentence-transformers
faiss-cpu
python-dotenv
groq